backend/requirements.txt
fastapi>=0.90.0
uvicorn>=0.20.0
pydantic>=1.10.0
scikit-learn>=1.0.0
pandas
requests>=2.28.0
python-dotenv>=0.21.0